added jwt and jwt authorization

+++ backend/controllers/goalController.js
@@ -1,9 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goal.model");
+const User = require("../models/user.model");
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.json({
     message: goals,
+    user: req.user.id,
   });
 });
 const addGoals = asyncHandler(async (req, res) => {
@@ -13,6 +15,7 @@ const addGoals = asyncHandler(async (req, res) => {
     throw new Error("please provide a text");
   }
   const goal = await Goal.create({
+    user: req.user.id,
     text,
   });
   res.json({
@@ -35,6 +38,16 @@ const updateGoals = asyncHandler(async (req, res) => {
       new: true,
     }
   );
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // make sure the logged in user can delete his goals
+  if (goal.user.id.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized to delete");
+  }
   res.json({
     message: updatedGoal,
   });
@@ -44,7 +57,17 @@ const deleteGoals = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(id);
   if (!goal) {
     res.status(404);
-    throw new Error({ message: "Goal not found" });
+    throw new Error("Goal not found");
+  }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // make sure the logged in user can delete his goals
+  if (goal.user.id.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized to delete");
   }
   const deletedGoal = await Goal.findByIdAndDelete(id);
   res.json({
